requirements.txt
django
markdown
python-dateutil
django-filter
social-auth-app-django
rest-social-auth
polib
python-dotenv